Reduced down to one client token, organized simplified code into new auth package
