+++ Dockerfile.dev
@@ -6,6 +6,5 @@ COPY go.sum go.sum
 COPY go.mod go.mod
 RUN go mod download
 RUN mkdir keys/
-RUN go run pkg/client/pem/main.go /app/keys terra-major-webapp
-RUN go run pkg/client/pem/main.go /app/keys unity-client
+RUN go run pkg/auth/pem/main.go /app/keys terra-major-client
 ENTRYPOINT go run main.go
