+++ pkg/characters/handlers/CreateCharacter.go
@@ -5,8 +5,8 @@ import (
 	"net/http"
 
 	accounts "github.com/mw-felker/terra-major-api/pkg/accounts/models"
+	authClient "github.com/mw-felker/terra-major-api/pkg/auth/client"
 	characters "github.com/mw-felker/terra-major-api/pkg/characters/models"
-	webAppClient "github.com/mw-felker/terra-major-api/pkg/client/webapp"
 	"github.com/mw-felker/terra-major-api/pkg/core"
 	sandboxes "github.com/mw-felker/terra-major-api/pkg/sandboxes/models"
 	utils "github.com/mw-felker/terra-major-api/pkg/utils"
@@ -14,7 +14,7 @@ import (
 
 func CreateCharacter(app *core.App) http.HandlerFunc {
 	return func(writer http.ResponseWriter, request *http.Request) {
-		claims, err := webAppClient.ParseAndValidateToken(request)
+		claims, err := authClient.ParseAndValidateToken(request)
 		if err != nil {
 			utils.ReturnError(writer, err.Error(), http.StatusUnauthorized)
 			return
